Refactor update_workspace_users into domain layer
