+++ authz/domain/workspace_users.py
@@ -1,6 +1,7 @@
 from sqlalchemy.orm.exc import NoResultFound
 
-from authz.models import User, WorkspaceRole
+from authz.database import db
+from authz.models import User, WorkspaceRole, Role
 from .exceptions import NotFoundError
 
 
@@ -27,6 +28,34 @@ class WorkspaceUser(object):
             workspace_role = None
 
         return cls(user, workspace_role)
+
+    @classmethod
+    def add_many(cls, workspace_id, workspace_user_dicts):
+        workspace_users = []
+
+        for user_dict in workspace_user_dicts:
+            try:
+                user = User.query.filter_by(id=user_dict["id"]).one()
+            except NoResultFound:
+                default_role = Role.query.filter_by(name="developer").one_or_none()
+                user = User(id=user_dict["id"], atat_role=default_role)
+
+            try:
+                role = Role.query.filter_by(name=user_dict["workspace_role"]).one()
+            except NoResultFound:
+                raise NotFoundError("role")
+
+            new_workspace_role = WorkspaceRole(user=user, role_id=role.id, workspace_id=workspace_id)
+            user.workspace_roles.append(new_workspace_role)
+
+            workspace_user = WorkspaceUser(user, new_workspace_role)
+            workspace_users.append(workspace_user)
+
+            db.session.add(user)
+
+        db.session.commit()
+
+        return workspace_users
 
     def permissions(self):
         atat_permissions = set(self.user.atat_role.permissions)
